refactor: 통계봇 Redis 의존성 제거
