+++ blink-batch/src/main/kotlin/com/jordyma/blink/domain/notification/dto/UserDataNotificationDto.kt
@@ -5,6 +5,7 @@ data class UserDataNotificationDto (
     val totalUserCount: Int = 0,
     val newUserFeed: Int = 0,
     val existingUserFeed: Int = 0,
+
     val linkViewCount: Int = 0,  // 일일 링크 조회 횟수
     val dailyActiveUsers: Int = 0 // 일일 활성 사용자수
 )
