+++ blink-core/src/main/kotlin/com.jordyma.blink/stats/service/LinkStatsProcessService.kt
@@ -2,11 +2,11 @@ package com.jordyma.blink.stats.service
 
 interface LinkStatsProcessService {
     // 일일 링크 조회수 저장
-    fun saveLinkViewCount(): Int
+    fun saveLinkViewCount(): Long
 
     // 일일 링크 조회수 조회
-    fun getYesterdayLinkViewCount(): Int
+    fun getYesterdayLinkViewCount(): Long
 
     // 일일 활성 사용자수 조회
-    fun getYesterdayDailyActiveUsers(): Int
+    fun getYesterdayDailyActiveUsers(): Long
 }
